# docs/openapi-gateway.yaml
openapi: 3.1.0
info:
  title: Wooblay Secure Gateway
  description: |
    Execute any action securely through Wooblay's three-layer security moat.
    Every action passes through policy evaluation, scope boundaries, pre-execution
    simulation, and ephemeral container execution. Credentials are never exposed —
    they're resolved from the vault at execution time and destroyed after.

    There is no fixed action list. Any action can be executed — provide the action
    name (for policy matching and audit), the command to run, and which provider's
    credentials to use. Wooblay handles the rest.

    **For OpenAI GPT Actions**: Import this spec as a custom action. Paste your
    Wooblay API key (starts with `wbl_ak_`) in the authentication section.

    **For Claude MCP / LangChain / any agent**: Use the Bearer token auth with
    your Wooblay API key.
  version: 1.0.0
  contact:
    name: Wooblay
    url: https://wooblay.com

servers:
  - url: https://gate.wooblay.com
    description: Production gateway

security:
  - apiKey: []

paths:
  /api/gateway/execute:
    post:
      operationId: executeAction
      summary: Execute a secure action
      description: |
        Execute any action through Wooblay's secure gateway. The action goes through:
        1. Policy evaluation (org-level rules)
        2. Scope boundary check (per-connection allowed/blocked targets)
        3. Pre-execution simulation (dry-run, API check, or evidence collection)
        4. Ephemeral secure execution (Docker container with injected credentials)

        The agent never sees the credentials. They're resolved from the vault,
        injected into an ephemeral container, and destroyed after execution.

        Any action name is accepted — there is no fixed list. The action name is
        used for policy matching and audit trail. Provide the command to run and
        which provider's credentials to inject.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
                - params
              properties:
                action:
                  type: string
                  description: |
                    Action name (any string). Used for policy matching and audit trail.
                    Examples: "deploy:staging", "db:migrate", "notification:send"
                params:
                  type: object
                  description: |
                    Parameters for the action. For generic execution, provide:
                    - command (string): The command to run in the ephemeral container
                    - provider (string): Which connection's credentials to inject
                    - image (string, optional): Docker image to use (default: node:20-slim)
                    - env (object, optional): Additional environment variables
                  additionalProperties: true
      responses:
        '200':
          description: Action executed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      stdout:
                        type: string
                        description: Standard output from the action
                      stderr:
                        type: string
                        description: Standard error from the action
                      exitCode:
                        type: integer
                        description: Exit code (0 = success)
                      containerId:
                        type: string
                        description: Ephemeral container ID (for audit)
                      durationMs:
                        type: integer
                        description: Execution duration in milliseconds
                      description:
                        type: string
                        description: Human-readable description of what happened
                  simulation:
                    type: object
                    nullable: true
                    properties:
                      strategy:
                        type: string
                        enum: [SANDBOX_EXEC, CONTENT_ANALYSIS]
                      passed:
                        type: boolean
                      summary:
                        type: string
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Authentication required
        '403':
          description: Denied by policy, scope boundary, or simulation failure
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                  riskTier:
                    type: string
                  category:
                    type: string
        '404':
          description: No active connection found for the specified provider
        '502':
          description: Secure execution failed

  /api/gateway/capabilities:
    get:
      operationId: getCapabilities
      summary: Get connected providers and usage info
      description: |
        Returns which providers the org has connected (and therefore which
        credentials are available for execution). Also returns usage instructions.
        There is no fixed action list — any action can be executed.
      responses:
        '200':
          description: Connected providers and usage info
          content:
            application/json:
              schema:
                type: object
                properties:
                  providers:
                    type: array
                    items:
                      type: object
                      properties:
                        provider:
                          type: string
                          description: Provider identifier
                        name:
                          type: string
                          description: Human-readable connection name
                  usage:
                    type: object
                    description: Instructions for calling the execute endpoint

components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: |
        Wooblay API key (starts with `wbl_ak_`).
        Create one at Settings > External API Keys in the Wooblay dashboard.